Translated fresnel
